Refactoring the DependencyInjectionNew.cs to DependencyInjection.cs Fixed tests and benchmarks. BugFix: `EventMappingOptionsBuilder` and `EventProjectionOptionsBuilder` now use `Assembly.GetEntryAssembly()` (not calling assembly)
