--- src/EventSourcing/DI/DependencyInjectionNew.cs
@@ -1,37 +1,0 @@
-
-
-using EventSourcing;
-using EventSourcing.Contexts;
-using EventSourcing.Mappers;
-using EventSourcing.Projections;
-using EventSourcing.Stores;
-// ReSharper disable once CheckNamespace
-using EventSourcing.Repositories;
-
-namespace Microsoft.Extensions.DependencyInjection;
-
-public static class DependencyInjectionNew
-{
-    /// <summary>
-    /// <para>
-    /// Registers the required services to use the event sourcing framework.
-    /// </para>
-    /// <para>
-    /// The following services are registered:<br/>
-    /// The <see cref="IEventRepository"/> interface for querying the event store - you will only work with that most of the time.<br/>
-    /// The <see cref="IEventBus"/> interface for publishing events - you can work with that, but you'll probably never do (the repository makes use of it to publish the events for projections etc.).<br/>
-    /// The <see cref="IEventStore"/> interface for persistence - it is only meant to be used internal by the repository.<br/>
-    /// The <see cref="IEventStoreDbContext"/> interface for the database context - you should leave that alone, but you can configure it in the <paramref name="options"/>.
-    /// </para>
-    /// </summary>
-    /// <param name="services">The <see cref="IServiceCollection"/> for the dependency injection.</param>
-    /// <param name="options">An action to configure the database with.</param>
-    /// <returns>The <see cref="IServiceCollection"/> for the dependency injection.</returns>
-    public static IServiceCollection AddEventSourcingNew(this IServiceCollection services, Action<EventSourcingOptionsBuilder>? options)
-    {
-        var builder = new EventSourcingOptionsBuilder(services);
-        options?.Invoke(builder);
-        builder.Build();
-        return services;
-    }
-}
