+++ tests/EventSourcing.FunctionTests.DI/DependencyInjectionTests.cs
@@ -1,6 +1,6 @@
 using System.Reflection;
-using EventSourcing.FunctionTests.DI.ValidAssembly.Events;
 using EventSourcing.Mappers;
+using EventSourcing.Projections;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 
@@ -9,25 +9,27 @@ namespace EventSourcing.FunctionTests.DI;
 public class DependencyInjectionTests
 {
     [Fact]
-    public void RegistrationShouldWork_WhenEventMappersAndEventsAreCorrect()
+    public void RegistrationShouldWork_WhenEventsAndMappersAndProjectionsAreCorrect()
     {
         var services = new ServiceCollection();
-        services.AddEventSourcing(builder => builder.UseInMemoryDatabase("Test"))
-            .AddEventMappers(config =>
-            {
-                var assembly = typeof(CustomEvent).Assembly;
-                config
-                    .AddCustomMappers(assembly)
-                    .AddDefaultMappers(assembly);
-            });
-        services.AddProjections(config => config.AddProjections(Assembly.GetExecutingAssembly()));
+        services.AddEventSourcing(options =>
+        {
+            options.ConfigureEventStoreDbContext(dbOptions => dbOptions.UseInMemoryDatabase("Test"));
+            options.ConfigureMapping(mappingOptions => mappingOptions.AddMappers(typeof(ValidAssembly.Events.CustomEvent).Assembly));
+            options.ConfigureProjections(projectionOptions => projectionOptions.AddProjections(typeof(ValidAssembly.Events.CustomEvent).Assembly));
+        });
         var provider = services.BuildServiceProvider();
 
         var eventMappers = provider.GetRequiredService<IEnumerable<IEventMapper>>().ToArray();
+        var projections = provider.GetRequiredService<IEnumerable<IEventHandler>>().Select(s => (IEventHandler)s).ToList();
         
-        eventMappers.Should().ContainSingle(m => m.Types.Contains("my-custom-event-v1") && m.EventType == typeof(CustomEvent));
-        eventMappers.Should().ContainSingle(m => m.Types.Contains("default-event-v1") && m.EventType == typeof(DefaultEvent));
+        eventMappers.Should().ContainSingle(m => m.Types.Contains("my-custom-event-v1") && m.EventType == typeof(ValidAssembly.Events.CustomEvent));
+        eventMappers.Should().ContainSingle(m => m.Types.Contains("default-event-v1") && m.EventType == typeof(ValidAssembly.Events.DefaultEvent));
         eventMappers.Should().HaveCount(2);
+        
+        projections.Should().Contain(p => p.GetType() == typeof(ValidAssembly.Projections.CustomEventProjection));
+        projections.Should().ContainSingle(p => p.GetType() == typeof(ValidAssembly.Projections.DefaultEventProjection));
+        projections.Should().HaveCount(3);
     }
     
     [Fact]
@@ -35,18 +37,14 @@ public class DependencyInjectionTests
     {
         var services = new ServiceCollection();
         
-        var func = () => services.AddEventSourcing(builder => builder.UseInMemoryDatabase("Test"))
-            .AddEventMappers(config =>
-            {
-                var assembly = typeof(InvalidAssembly.Events.CustomEvent).Assembly;
-                config
-                    .AddCustomMappers(assembly)
-                    .AddDefaultMappers(assembly);
-            });
+        var func = () => services.AddEventSourcing(options =>
+        {
+            options.ConfigureEventStoreDbContext(dbOptions => dbOptions.UseInMemoryDatabase("Test"));
+            options.ConfigureMapping(mappingOptions => mappingOptions.AddMappers(typeof(InvalidAssembly.Events.CustomEvent).Assembly));
+            options.ConfigureProjections(projectionOptions => projectionOptions.AddProjections(Assembly.GetExecutingAssembly()));
+        });
 
         func.Should().Throw<InvalidOperationException>().WithMessage("*multiple*");
+
     }
 }
-
-
-
