+++ tests/EventSourcing.FunctionTests/Mappers/EventRegistryTests.cs
@@ -15,7 +15,11 @@ public class EventRegistryTests
     public void DependencyInjection_ShouldRegisterAbstractEventMappersInAssembly()
     {
         var services = new ServiceCollection();
-        services.AddEventMappers(config => config.AddCustomMapper<MagicEventMapper>());
+        services.AddEventSourcing(config =>
+        {
+            config.ConfigureMapping(options => options.AddMapper<MagicEventMapper>());
+            config.ConfigureProjections(options => options.IgnoreUncoveredEvents());
+        });
 
         services.Should().ContainSingle(s => s.ImplementationType == typeof(MagicEventMapper));
         
@@ -29,7 +33,11 @@ public class EventRegistryTests
     public void SerializeAndDeserialize_ShouldSucceed_ForNewestVersion()
     {
         var services = new ServiceCollection();
-        services.AddEventMappers(config => config.AddCustomMapper<MagicEventMapper>());
+        services.AddEventSourcing(config =>
+        {
+            config.ConfigureMapping(options => options.AddMapper<MagicEventMapper>());
+            config.ConfigureProjections(options => options.IgnoreUncoveredEvents());
+        });
         var serviceProvider = services.BuildServiceProvider();
         var registry = serviceProvider.GetRequiredService<IEventRegistry>();
         var magicEvent = new MagicEvent(Guid.NewGuid(), "Magic", DateTime.UtcNow);
@@ -50,8 +58,11 @@ public class EventRegistryTests
     public void Deserialize_ShouldSucceed_ForOldVersion2()
     {
         var services = new ServiceCollection();
-        
-        services.AddEventMappers(config => config.AddCustomMapper<MagicEventMapper>());
+        services.AddEventSourcing(config =>
+        {
+            config.ConfigureMapping(options => options.AddMapper<MagicEventMapper>());
+            config.ConfigureProjections(options => options.IgnoreUncoveredEvents());
+        });
         var serviceProvider = services.BuildServiceProvider();
         var registry = serviceProvider.GetRequiredService<IEventRegistry>();
         var magicEventV2 = new MagicEventMapper.MagicEventV2(Guid.NewGuid(), "Magic", DateTime.UtcNow);
@@ -69,7 +80,11 @@ public class EventRegistryTests
     public void Deserialize_ShouldSucceed_ForOldVersion1()
     {
         var services = new ServiceCollection();
-        services.AddEventMappers(config => config.AddCustomMapper<MagicEventMapper>());
+        services.AddEventSourcing(config =>
+        {
+            config.ConfigureMapping(options => options.AddMapper<MagicEventMapper>());
+            config.ConfigureProjections(options => options.IgnoreUncoveredEvents());
+        });
         var serviceProvider = services.BuildServiceProvider();
         var registry = serviceProvider.GetRequiredService<IEventRegistry>();
         var magicEventV1 = new MagicEventMapper.MagicEventV1(Guid.NewGuid(), DateTime.UtcNow);
@@ -87,7 +102,11 @@ public class EventRegistryTests
     public void Deserialize_ShouldThrowException_WhenMapperNotFound()
     {
         var services = new ServiceCollection();
-        services.AddEventMappers(config => config.AddCustomMapper<MagicEventMapper>());
+        services.AddEventSourcing(config =>
+        {
+            config.ConfigureMapping(options => options.AddMapper<MagicEventMapper>());
+            config.ConfigureProjections(options => options.IgnoreUncoveredEvents());
+        });
         var serviceProvider = services.BuildServiceProvider();
         var registry = serviceProvider.GetRequiredService<IEventRegistry>();
         var magicEvent = new MagicEvent(Guid.NewGuid(), "Magic", DateTime.UtcNow);
@@ -102,7 +121,11 @@ public class EventRegistryTests
     public void Serialize_ShouldThrowException_WhenMapperNotFound()
     {
         var services = new ServiceCollection();
-        services.AddEventMappers(config => config.AddCustomMapper<MagicEventMapper>());
+        services.AddEventSourcing(config =>
+        {
+            config.ConfigureMapping(options => options.AddMapper<MagicEventMapper>());
+            config.ConfigureProjections(options => options.IgnoreUncoveredEvents());
+        });
         var serviceProvider = services.BuildServiceProvider();
         var registry = serviceProvider.GetRequiredService<IEventRegistry>();
